Created the plugin
Its pretty cool honestly. When the server is shut down, it saves the players inventories in files. When the server is started everything is loaded into a hashmap. If there was a block there (like a pressure plate), the block will be placed there after the chest disappears when the player joins again. If the spot is occupied by another block, it will simply drop as an item.
Disclaimer - I have not tested/fixed the following:
- Multiple players access and edit the inventory at the same time (maybe item duping possible)
- There might be issues with chests being placed where a DisconnectChest has previously stood, although that should be taken care of
- Newest update of the saving-previous-block feature has not been tested, it should work though